Feat: remember camera position when returning from restaurant detail
